backend/pyproject.toml
[build-system]
requires = ["uv_build>=0.7.5,<0.8"]
build-backend = "uv_build"

[project]
name = "backend"
version = "0.1.0"
description = "TanStack AI + ADK HITL Demo Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "asyncpg",
    "pandas",
    "google-adk",
    "structlog",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "ty",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.ruff]
target-version = "py312"
exclude = [".venv", "alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ty.src]
include = ["src"]
exclude = [".venv/", "alembic", "tests"]
